updating tests, adding some webdriver flags
