+++ vector_search/conftest.py
@@ -22,6 +22,40 @@ class DummyEmbedEncoder(BaseEncoder):
 
 
 @pytest.fixture(autouse=True)
+def marketing_metadata_mocks(mocker):
+    mocker.patch(
+        "vector_search.utils._fetch_page",
+        return_value="""
+        <html>
+        <body>
+            <div class="container">
+            <div class="learning-header">
+              <h1>WHAT YOU WILL LEARN</h1>
+              <p data-block-key="fq16h">MIT xPRO is collaborating with online
+              education provider Emeritus to deliver this online program.</p>
+            </div>
+            <ul class="learning-outcomes-list d-flex flex-wrap justify-content-between">
+              <li>Learn to code in Python</li>
+              <li>Use SQL to create databases</li>
+              <li>Wrangle and analyze millions of pieces of
+              data using databases in Python</li>
+              <li>Understand how networks work, including IPs,
+              security, and servers</li>
+              <li>Manage big data using data warehousing and
+              workflow management platforms</li>
+              <li>Use cutting-edge data engineering
+              platforms and tools to manage data</li>
+              <li>Explore artificial intelligence
+              and machine learning concepts,
+              including reinforcement learning and deep neural networks</li>
+            </ul>
+          </div>
+        </body>
+        </html>""",
+    )
+
+
+@pytest.fixture(autouse=True)
 def _use_dummy_encoder(settings):
     settings.QDRANT_ENCODER = "vector_search.conftest.DummyEmbedEncoder"
     settings.QDRANT_DENSE_MODEL = None
