+++ vector_search/utils.py
@@ -611,11 +611,16 @@ def html_to_markdown(html):
 
 @cache
 def _get_web_driver():
-    chrome_options = Options()
-    chrome_options.add_argument("--headless")
     service = webdriver.ChromeService(executable_path=which("chromedriver"))
+    chrome_options = Options()
+    chrome_options.add_argument("--headless=new")
     chrome_options.add_argument("--no-sandbox")
     chrome_options.add_argument("--disable-gpu")
+    chrome_options.add_argument("--disable-extensions")
+    chrome_options.add_argument("--disable-features=site-per-process")
+    chrome_options.add_argument("--incognito")
+    chrome_options.add_argument("--disable-dev-shm-usage")
+
     return webdriver.Chrome(service=service, options=chrome_options)
